requirements.txt
pip~=24.3.1
attrs~=24.2.0
pillow~=11.0.0
tornado~=6.4.2
Jinja2~=3.1.4
cvzone~=1.6.1
opencv-python~=4.10.0.84
mediapipe~=0.10.18
numpy~=1.26.4
Flask~=3.1.0